APM: move mock fetchers and queries out of api-core/api-queries
Mock data generation no longer ships in the api-core package; it lives
in client/dashboard/sites/performance/backend/mock-data.ts alongside the
queryOptions factories that wrap it. api-core retains only the APM
types; api-queries retains only the real enable/disable mutation. The
backend index and request-detail routes now prefetch via loaders, and
the relative-time and millisecond strings in slow-requests-list /
request-detail go through sprintf + __ for i18n.
