+++ client/dashboard/sites/performance/backend/request-detail/index.tsx
@@ -1,21 +1,26 @@
-import { siteApmRequestQuery, siteBySlugQuery } from '@automattic/api-queries';
+import { siteBySlugQuery } from '@automattic/api-queries';
 import { useSuspenseQuery } from '@tanstack/react-query';
 import {
 	__experimentalHStack as HStack,
 	__experimentalText as Text,
 	__experimentalVStack as VStack,
 } from '@wordpress/components';
-import { __ } from '@wordpress/i18n';
+import { __, sprintf } from '@wordpress/i18n';
 import Breadcrumbs from '../../../../app/breadcrumbs';
 import { Card, CardHeader } from '../../../../components/card';
 import { PageHeader } from '../../../../components/page-header';
 import PageLayout from '../../../../components/page-layout';
+import { siteApmRequestQuery } from '../mock-data';
 
 function formatMs( ms: number ): string {
 	if ( ms >= 1000 ) {
 		return `${ ( ms / 1000 ).toFixed( 2 ) } s`;
 	}
-	return `${ ms } ms`;
+	return sprintf(
+		/* translators: %d is a number of milliseconds. */
+		__( '%d ms' ),
+		ms
+	);
 }
 
 export default function RequestDetail( {
