+++ packages/api-queries/src/site-apm.ts
@@ -1,10 +1,5 @@
-import {
-	fetchApmOverview,
-	fetchApmRequest,
-	fetchApmSlowRequests,
-	updateApmEnabled,
-} from '@automattic/api-core';
-import { mutationOptions, queryOptions } from '@tanstack/react-query';
+import { updateApmEnabled } from '@automattic/api-core';
+import { mutationOptions } from '@tanstack/react-query';
 import { queryClient } from './query-client';
 import { siteQueryFilter } from './site';
 import type { Site } from '@automattic/api-core';
@@ -22,22 +17,4 @@ export const siteApmEnabledMutation = ( siteId: number ) =>
 					: site
 			);
 		},
-	} );
-
-export const siteApmOverviewQuery = ( siteId: number ) =>
-	queryOptions( {
-		queryKey: [ 'site', siteId, 'apm', 'overview' ],
-		queryFn: () => fetchApmOverview( siteId ),
-	} );
-
-export const siteApmSlowRequestsQuery = ( siteId: number ) =>
-	queryOptions( {
-		queryKey: [ 'site', siteId, 'apm', 'slow-requests' ],
-		queryFn: () => fetchApmSlowRequests( siteId ),
-	} );
-
-export const siteApmRequestQuery = ( siteId: number, requestId: string ) =>
-	queryOptions( {
-		queryKey: [ 'site', siteId, 'apm', 'request', requestId ],
-		queryFn: () => fetchApmRequest( siteId, requestId ),
 	} );
